requirements.txt
# transformers==4.51.3
tqdm==4.67.1
accelerate==1.7.0
datasets==3.6.0
decord==0.6.0
flash_attn==2.7.4.post1
google-genai==1.16.1
librosa==0.11.0
loguru==0.7.3
numpy==2.1.2
omegaconf==2.3.0
openai==1.79.0
peft==0.15.2
qwen-omni-utils==0.0.4
soundfile==0.13.1
tokenizers==0.21.1
ujson==5.10.0
sentencepiece==0.2.0